Add local queue system
